Working force field, but seems like a dead end
To get anything close to the quality of the existing collision checking, the force field needs to be high resolution, and a high resolution FF is even worse performance. It may be I could get something better with bicubic interpolation or adaptive grid density. But this basic implementation is pretty disappointing. Still, coded it into the spec as an option for now.